Fix all rubocop offenses
