+++ .rubocop.yml
@@ -5,3 +5,10 @@ AllCops:
     - 'db/seeds.rb'
     - 'node_modules/**/*'
     - 'bin/*'
+
+Metrics/BlockLength:
+  Exclude:
+    - 'spec/**/*_spec.rb'
+    - 'config/routes.rb'
+    - 'config/environments/development.rb'
+    - 'lib/tasks/youtube_seed.rake'
